postMapper class 추가 및 Mapper 추가로 인한 중복 코드 제거
