+++ src/main/java/com/team/neorangloa/domain/post/dto/PostResponse.java
@@ -1,12 +1,10 @@
 package com.team.neorangloa.domain.post.dto;
 
 import com.team.neorangloa.domain.post.entity.Post;
-import com.team.neorangloa.domain.user.entity.User;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
-import javax.validation.constraints.NotEmpty;
 
 @Getter
 @Builder
@@ -21,14 +19,4 @@ public class PostResponse {
     private final String content;
     private final String image;
 
-    @Builder
-    public static PostResponse of(Post post) {
-        return PostResponse.builder()
-                .postId(post.getId())
-                .nickname(post.getAuthor().getNickname())
-                .title(post.getTitle())
-                .content(post.getContent())
-                .image(post.getPost_image())
-                .build();
-    }
 }
